fix(structure): moved markdown component, refactored services and pipes.
