+++ src/app/iteration/iteration.module.ts
@@ -13,7 +13,7 @@ import {
 import { FabPlannerIterationModalComponent } from './iteration-modal/iteration-modal.component';
 import { MyDatePickerModule } from 'mydatepicker';
 import { IterationComponent } from './iteration.component';
-import { IterationService } from './iteration.service';
+import { IterationService } from '../../services/iteration.service';
 import { ModalModule } from 'ngx-modal';
 import { TooltipModule } from 'ng2-bootstrap';
 import { TruncateModule } from 'ng2-truncate';
