+++ src/app/login/login.component.ts
@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 import { AuthenticationService } from 'ngx-login-client';
 
 import { LoginItem } from './login-item';
-import { LoginService } from './login.service';
+import { LoginService } from '../services/login.service';
 
 
 @Component({
