+++ src/app/services/collaborator.service.ts
@@ -3,7 +3,7 @@ import { Observable } from 'rxjs/Observable';
 import { User, UserService } from 'ngx-login-client';
 import { Spaces } from 'ngx-fabric8-wit';
 import { Logger } from 'ngx-base';
-import { HttpService } from './../shared/http-service';
+import { HttpService } from '../shared/http-service';
 
 
 @Injectable()
