+++ src/app/services/login.service.ts
@@ -4,7 +4,7 @@ import { WIT_API_URL } from 'ngx-fabric8-wit';
 
 import 'rxjs/add/operator/toPromise';
 
-import { LoginItem } from './login-item';
+import { LoginItem } from '../login/login-item';
 
 @Injectable()
 export class LoginService {
