+++ src/app/work-item/work-item-detail-add-type-selector/work-item-detail-add-type-selector-widget/work-item-detail-add-type-selector-widget.component.ts
@@ -7,7 +7,7 @@ import { AuthenticationService } from 'ngx-login-client';
 import { Subscription } from 'rxjs/Subscription';
 import { Space, Spaces } from 'ngx-fabric8-wit';
 
-import { WorkItemService } from './../../work-item.service';
+import { WorkItemService } from '../../../services/work-item.service';
 import { WorkItemListEntryComponent } from './../../work-item-list/work-item-list-entry/work-item-list-entry.component';
 import { WorkItemType } from './../../../models/work-item-type';
 import { WorkItem } from './../../../models/work-item';
