+++ src/app/work-item/work-item-detail/work-item-comment/work-item-comment.component.spec.ts
@@ -48,25 +48,25 @@ import {
 import { ModalModule } from 'ngx-modal';
 
 import { AreaModel } from '../../../models/area.model';
-import { AreaService } from '../../../area/area.service';
-import { CollaboratorService } from './../../../collaborator/collaborator.service';
+import { AreaService } from '../../../services/area.service';
+import { CollaboratorService } from '../../../services/collaborator.service';
 import { DynamicFieldComponent } from './../dynamic-form/dynamic-field.component';
 import { TypeaheadDropdown } from './../typeahead-dropdown/typeahead-dropdown.component';
 import { IterationModel } from '../../../models/iteration.model';
-import { IterationService } from '../../../iteration/iteration.service';
-import { MarkdownControlComponent } from './../markdown-control/markdown-control.component';
+import { IterationService } from '../../../../services/iteration.service';
+import { MarkdownControlComponent } from '../../../markdown-control/markdown-control.component';
 import { LinkType } from '../../../models/link-type';
 import { Comment } from '../../../models/comment';
 import { WorkItem } from '../../../models/work-item';
 import { WorkItemType } from '../../../models/work-item-type';
 import { AlmUserName } from '../../../pipes/alm-user-name.pipe';
 import { SpacesService } from '../../../shared/standalone/spaces.service';
-import { WorkItemService } from '../../work-item.service';
+import { WorkItemService } from '../../../services/work-item.service';
 import { WorkItemLinkComponent } from '../work-item-link/work-item-link.component';
 import { WorkItemCommentComponent } from './work-item-comment.component';
 import { WorkItemDetailComponent } from '../work-item-detail.component';
-import { WorkItemLinkTypeFilterByTypeName, WorkItemLinkFilterByTypeName } from '../work-item-detail-pipes/work-item-link-filters.pipe';
-import { WorkItemTypeControlService } from './../../work-item-type-control.service';
+import { WorkItemLinkTypeFilterByTypeName, WorkItemLinkFilterByTypeName } from '../../../pipes/work-item-link-filters.pipe';
+import { WorkItemTypeControlService } from '../../../services/work-item-type-control.service';
 
 describe('Comment section for the work item detailed view - ', () => {
   let comp: WorkItemDetailComponent;
