+++ src/app/work-item/work-item-list/work-item-list.component.spec.ts
@@ -35,10 +35,10 @@ import { WorkItemQuickAddModule } from '../work-item-quick-add/work-item-quick-a
 
 import { WorkItem } from '../../models/work-item';
 import { WorkItemType } from '../../models/work-item-type';
-import { WorkItemService } from '../work-item.service';
+import { WorkItemService } from '../../services/work-item.service';
 
 import { IterationModel } from '../../models/iteration.model';
-import { IterationService } from '../../iteration/iteration.service';
+import { IterationService } from '../../../services/iteration.service';
 
 import { WorkItemListEntryComponent } from './work-item-list-entry/work-item-list-entry.component';
 import { WorkItemListComponent } from './work-item-list.component';
